taisyti

+++ PAPAS/src/main/java/lt/pap/dao/TofModelRepository.java
@@ -4,7 +4,7 @@ import lt.pap.model.TofModel;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TofModelRepository extends JpaRepository<TofModel, Long>
+public interface TofModelRepository extends JpaRepository<TofModel, Integer>
 {
     
 }
